feat: ajustar tipos de datos, agregar enums y estado en migraciones

+++ database/migrations/2026_05_30_123639_create_carrera_gestions_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('carrera_gestion', function (Blueprint $table) {
             $table->string('codigo_carrera', 20);
             $table->string('codigo_gestion', 20);
-            $table->integer('cupos')->default(0);
+            $table->unsignedSmallInteger('cupos')->default(0);
             
             $table->primary(['codigo_carrera', 'codigo_gestion']);
